+++ bin/runner.js
@@ -50,9 +50,9 @@ function generate(model, allTestsGenerated) {
         });
       });
     } else {
-      var pathToModule = path.join(
-        model.testsPath,
-        model.root,
+      var pathToModule = findModule(
+        path.join(model.testsPath, ".."),
+        model["source-directories"],
         elmModuleToPath(inputName)
       );
 
@@ -73,10 +73,17 @@ function generate(model, allTestsGenerated) {
   });
 
   var writtenTests = 0;
+  var noExamples = 0;
+  app.ports.noExamples.subscribe(function () {
+    noExamples = noExamples + 1;
+  });
   app.ports.writeFiles.subscribe(function (data) {
     serial(data, writeFile(model.testsDocPath), function () {
       writtenTests = writtenTests + 1;
-      if (writtenTests === model.tests.length && allTestsGenerated) {
+      if (
+        writtenTests + noExamples === model.tests.length &&
+        allTestsGenerated
+      ) {
         allTestsGenerated(warnings);
       }
     });
@@ -154,7 +161,7 @@ function forFiles(model, files) {
     .filter(function (v) {
       return v.endsWith(".elm");
     })
-    .map(elmPathToModule(model.root, model.testsPath));
+    .map(elmPathToModule("..", model["source-directories"], model.testsPath));
 
   return model;
 }
@@ -215,17 +222,28 @@ function writeFile(testsDocPath) {
   };
 }
 
-function elmPathToModule(root, testsPath) {
+function elmPathToModule(elmRoot, sourceDirs, testsPath) {
   return function (pathName) {
-    var relativePath = path.relative(
-      path.resolve(path.join(testsPath, root)),
-      pathName
-    );
+    var relativePath = findModule(elmRoot, sourceDirs, pathName);
     if (relativePath.startsWith("./")) {
       relativePath = relativePath.substr(2);
     }
     return relativePath.substr(0, relativePath.length - 4).replace(/\//g, ".");
   };
+}
+
+function findModule(elmRoot, sourceDirs, pathName) {
+  for (var i = 0; i < sourceDirs.length; i++) {
+    const dir = sourceDirs[i];
+    const module = path.join(elmRoot, dir, pathName);
+    if (fs.existsSync(module)) {
+      return module;
+    }
+  }
+  console.error(
+    `Could not find module ${pathName} in ${sourceDirs} with root ${elmRoot}`
+  );
+  process.exit(1);
 }
 
 function elmModuleToPath(moduleName) {